Move ForgotPassword to containers
