+++ test/pages/forgot-password.js
@@ -4,7 +4,7 @@ import { expect } from 'chai'
 import { mount } from 'enzyme'
 
 import Config from '../../src/config.js'
-import ForgotPassword from '../../src/pages/forgot-password'
+import ForgotPassword from '../../src/containers/forgot-password'
 import ForgotPasswordForm from '../../src/components/forgot-password-form'
 
 const BASE_URI = 'http://localhost:8080'
